+++ git/src/lib.rs
@@ -34,11 +34,11 @@ pub use extension::{ExtensionConfig, list_extensions};
 pub use fetch_prune::FetchPruneConfig;
 pub use http_options::HttpOptions;
 pub use path::{git_dir, lfs_alternate_dirs, lfs_dir, work_tree_root};
-pub use refs::{RecentRef, RefKind, recent_branches};
+pub use refs::{RecentRef, RefKind, WorktreeEntry, recent_branches, worktrees};
 pub use rev_list::{RevListEntry, rev_list, rev_list_with_args};
 pub use scanner::{
-    PointerEntry, TreeBlob, scan_index_lfs, scan_pointers, scan_pointers_with_args,
-    scan_previous_versions, scan_tree, scan_tree_blobs,
+    PointerEntry, TreeBlob, scan_index_lfs, scan_index_pointers, scan_pointers,
+    scan_pointers_with_args, scan_previous_versions, scan_stashed, scan_tree, scan_tree_blobs,
 };
 
 #[derive(Debug, thiserror::Error)]